* Components/Searching/NSSet+ArticleFiltering.m         * Components/ArticleTable/ArticleTablePlugin.m           some NSInteger conversions, and clang warning shutup
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2223 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/ArticleTable/ArticleTablePlugin.m
@@ -212,7 +212,7 @@ int compareArticleRatings( id articleA, id articleB, void* context) {
 -(void) tableView: (NSTableView*) aTableView
     willDisplayCell: (id)aCell
     forTableColumn: (NSTableColumn*) aTableColumn
-    row: (int)rowIndex
+    row: (NSInteger)rowIndex
 {
     NSCell* cell = aCell;
     id<Article> article = [articles objectAtIndex: rowIndex];
@@ -246,7 +246,7 @@ int compareArticleRatings( id articleA, id articleB, void* context) {
 -(void) tableView: (NSTableView*) aTableView
    setObjectValue: (id) anObj
    forTableColumn: (NSTableColumn*) aTableColumn
-              row: (int) rowIndex
+              row: (NSInteger) rowIndex
 {
     if (aTableColumn == ratingCol) {
 	id article;
